+++ libs/qwt/src/qwt_plot_barchart.h
@@ -87,7 +87,7 @@ class QWT_EXPORT QwtPlotBarChart
 
     virtual void drawSeries( QPainter*,
         const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-        const QRectF& canvasRect, int from, int to ) const QWT_OVERRIDE;
+        const QRectF& canvasRect, size_t from, size_t to ) const QWT_OVERRIDE;
 
     virtual QRectF boundingRect() const QWT_OVERRIDE;
