+++ libs/qwt/src/qwt_plot_curve.h
@@ -299,9 +299,9 @@ class QWT_EXPORT QwtPlotCurve
     void setCurveFitter( QwtCurveFitter* );
     QwtCurveFitter* curveFitter() const;
 
-    virtual void drawSeries( QPainter*,
+    virtual void drawSeries(QPainter*,
         const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-        const QRectF& canvasRect, int from, int to ) const QWT_OVERRIDE;
+        const QRectF& canvasRect, size_t from, size_t to ) const QWT_OVERRIDE;
 
     virtual QwtGraphic legendIcon( int index, const QSizeF& ) const QWT_OVERRIDE;
