+++ libs/qwt/src/qwt_plot_histogram.cpp
@@ -331,9 +331,9 @@ void QwtPlotHistogram::setSamples(
 
    \sa drawOutline(), drawLines(), drawColumns
  */
-void QwtPlotHistogram::drawSeries( QPainter* painter,
+void QwtPlotHistogram::drawSeries(QPainter* painter,
     const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-    const QRectF& canvasRect, int from, int to ) const
+    const QRectF& canvasRect, size_t from, size_t to ) const
 {
     Q_UNUSED( canvasRect )
