+++ libs/qwt/src/qwt_plot_intervalcurve.cpp
@@ -319,9 +319,9 @@ QRectF QwtPlotIntervalCurve::boundingRect() const
 
    \sa drawTube(), drawSymbols()
  */
-void QwtPlotIntervalCurve::drawSeries( QPainter* painter,
+void QwtPlotIntervalCurve::drawSeries(QPainter* painter,
     const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-    const QRectF& canvasRect, int from, int to ) const
+    const QRectF& canvasRect, size_t from, size_t to ) const
 {
     if ( to < 0 )
         to = dataSize() - 1;
