+++ libs/qwt/src/qwt_plot_multi_barchart.cpp
@@ -371,9 +371,9 @@ QRectF QwtPlotMultiBarChart::boundingRect() const
 
    \sa drawSymbols()
  */
-void QwtPlotMultiBarChart::drawSeries( QPainter* painter,
+void QwtPlotMultiBarChart::drawSeries(QPainter* painter,
     const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-    const QRectF& canvasRect, int from, int to ) const
+    const QRectF& canvasRect, size_t from, size_t to ) const
 {
     if ( to < 0 )
         to = dataSize() - 1;
